requirements.txt
librosa~=0.9.2
essentia~=2.1b6.dev858
chumpy~=0.70

numpy~=1.23.4
scipy~=1.9.3
opencv-python~=4.6.0.66
pillow~=9.2.0
tqdm~=4.64.1
smplx~=0.1.26
matplotlib~=3.6.2
soundfile~=0.11.0
yaml~=0.2.5
pyyaml~=6.0
torchsummary~=1.5.1
easydict~=1.10
ffmpeg~=4.3
aniposelib~=0.3.9
imageio~=2.22.4
setuptools~=65.5.0